+++ moira_client/models/event.py
@@ -1,3 +1,4 @@
+from ..client import InvalidJSONError
 from ..client import ResponseStructureError
 
 
@@ -10,7 +11,7 @@ class EventManager:
 
     def fetch_by_trigger(self, trigger, limit=MAX_FETCH_LIMIT):
         """
-        Get all events by trigger.
+        Get all events by trigger
         :param trigger: Trigger trigger
         :param limit: int limit
         :return: list of dicts
@@ -29,6 +30,20 @@ class EventManager:
             raise ResponseStructureError("list doesn't exist in response", result)
 
         return result['list']
+
+    def delete_all(self):
+        """
+        Remove all events
+
+        :return: True on success, False otherwise
+        """
+        try:
+            result = self._client.delete(self._full_path("/all"))
+            return False
+        except InvalidJSONError as e:
+            if e.content == b'':  # successfully if response is blank
+                return True
+            return False
 
     def _full_path(self, path=''):
         if path: