+++ tests/models/test_notification.py
@@ -1,9 +1,12 @@
 try:
+    from unittest.mock import Mock
     from unittest.mock import patch
-except:
+except ImportError:
+    from mock import Mock
     from mock import patch
 
 from moira_client.client import Client
+from moira_client.client import InvalidJSONError
 from moira_client.client import ResponseStructureError
 from moira_client.models.notification import NotificationManager
 from .test_model import ModelTest
@@ -35,3 +38,25 @@ class NotificationTest(ModelTest):
 
         self.assertTrue(get_mock.called)
         get_mock.assert_called_with('notification', params=params)
+
+    def test_delete_all(self):
+        client = Client(self.api_url)
+        notification_manager = NotificationManager(client)
+
+        with patch.object(client, 'delete', new=Mock(side_effect=InvalidJSONError(b''))) as delete_mock:
+            res = notification_manager.delete_all()
+
+        self.assertTrue(delete_mock.called)
+        self.assertTrue(res)
+        delete_mock.assert_called_with('notification/all')
+
+    def test_delete_all_bad_response(self):
+        client = Client(self.api_url)
+        notification_manager = NotificationManager(client)
+
+        with patch.object(client, 'delete') as delete_mock:
+            res = notification_manager.delete_all()
+
+        self.assertTrue(delete_mock.called)
+        self.assertFalse(res)
+        delete_mock.assert_called_with('notification/all')